CI: Enable building of ARM / ARM64 in continuous integration and upload artifacts
We need to set DETOURS_TARGET_PROCESSOR to tell the makefile's what
architecture we are targeting. It seems like we should eventually
fix system.mak to sniff target architecture using %VSCMD_ARG_TGT_ARCH%.

This change also supports uploading the artifacts from the pipeline so
they are available for download and consumption.